requirements.txt
# Hugging Face Spaces installs from this file.
# Note: Local workspace packages (pdfsys-*) are loaded via sys.path in demo/app.py
# and do not need editable installation in HF Spaces.

# --- Python 3.13 compatibility (audioop removed) --------------------------
audioop-lts

# --- CPU-only torch (HF Spaces free tier is CPU) --------------------------
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1,<3.0

# --- Third-party runtime deps -------------------------------------------
gradio==5.12.0
huggingface-hub>=0.26,<0.29
pymupdf>=1.24
xgboost>=2.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.26
transformers>=4.44
pillow>=10.0
